Fix buffer concatenation
If you concatenated a Uint16Array, you'd only have half the bytes
(amplified if you use 32 byte views). This was because `TypedArray#set`
doesn't respect the byte size of each member, merging each index as the
literal value regardless of whether it fit.

Discovered since the json data is encoded in Uint16Arrays.