Fix node-style buffer conversion
If a `Buffer` instance was passed, it was being stringified with the
rest of the data and parsed back out. The tests missed it because I was
using `Buffer.from` on the result data, which detects it was once a
buffer and re-incorporates all the data. While it works, it's not as
space efficient as a pure binary array.